+++ projects/plots/plots/core/constants.py
@@ -1,0 +1,2 @@
+SECONDS_PER_MONTH = 60 * 60 * 24 * 30
+SECONDS_PER_YEAR = 60 * 60 * 24 * 365.25
